site/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1
django-admin-interface==0.28.8
django-cleanup==8.1.0
django-colorfield==0.11.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==74.0.0
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
uWSGI==2.0.21